# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: IPL Auction Simulator API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /seasons/{year}/teams:
    get:
      operationId: getSeasonTeams
      tags: [seasons]
      summary: Get teams for a season
      parameters:
        - name: year
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Teams for the season
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Franchise"
...
  /rooms/{code}/teams:
    get:
      operationId: getRoomTeams
      tags: [teams]
      summary: Get all teams in the room
      parameters:
        - name: code
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Teams in room
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Team"
  /rooms/{code}/teams/{franchiseId}/select:
    post:
      operationId: selectFranchise
      tags: [teams]
      summary: Select a franchise for the room
      parameters:
        - name: code
          in: path
          required: true
          schema:
            type: string
        - name: franchiseId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Franchise selected
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Team"
  /rooms/{code}/teams/{teamId}/retain:
    post:
      operationId: retainPlayers
      tags: [teams]
      summary: Submit player retentions for team
      parameters:
        - name: code
          in: path
          required: true
          schema:
            type: string
        - name: teamId
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/RetentionInput"
      responses:
        "200":
          description: Players retained
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Team"
  /rooms/{code}/teams/{teamId}/squad:
    get:
      operationId: getTeamSquad
      tags: [teams]
      summary: Get full team squad with stats
      parameters:
        - name: code
          in: path
          required: true
          schema:
            type: string
        - name: teamId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Team squad
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/TeamSquad"

  # Auction Pool
...
components:
  schemas:
...
    Franchise:
      type: object
      properties:
        id:
          type: integer
        name:
          type: string
        shortName:
          type: string
        city:
          type: string
        primaryColor:
          type: string
        secondaryColor:
          type: string
        logoUrl:
          type: ["string", "null"]
      required: [id, name, shortName, city, primaryColor, secondaryColor]
...
    Team:
      type: object
      properties:
        id:
          type: integer
        roomId:
          type: integer
        franchiseId:
          type: integer
        franchiseName:
          type: string
        shortName:
          type: string
        primaryColor:
          type: string
        secondaryColor:
          type: string
        budgetRemainingCrore:
          type: number
        budgetSpentCrore:
          type: number
        retentionComplete:
          type: boolean
        owners:
          type: array
          items:
            $ref: "#/components/schemas/TeamOwner"
      required: [id, roomId, franchiseId, franchiseName, shortName, primaryColor, secondaryColor, budgetRemainingCrore, budgetSpentCrore, retentionComplete, owners]
    TeamOwner:
      type: object
      properties:
        userId:
          type: string
        displayName:
          type: string
        isOnline:
          type: boolean
      required: [userId, displayName, isOnline]
    RetentionInput:
      type: object
      properties:
        playerIds:
          type: array
          items:
            type: integer
        retentionPricesCrore:
          type: array
          items:
            type: number
      required: [playerIds, retentionPricesCrore]
    TeamSquad:
      type: object
      properties:
        team:
          $ref: "#/components/schemas/Team"
        retainedPlayers:
          type: array
          items:
            $ref: "#/components/schemas/SquadPlayer"
        auctionedPlayers:
          type: array
          items:
            $ref: "#/components/schemas/SquadPlayer"
        batters:
          type: integer
        bowlers:
          type: integer
        allRounders:
          type: integer
        wicketKeepers:
          type: integer
        overseas:
          type: integer
        indian:
          type: integer
        averageAge:
          type: ["number", "null"]
      required: [team, retainedPlayers, auctionedPlayers, batters, bowlers, allRounders, wicketKeepers, overseas, indian]
    SquadPlayer:
      type: object
      properties:
        id:
          type: integer
        name:
          type: string
        role:
          type: string
        nationality:
          type: string
        isOverseas:
          type: boolean
        priceCrore:
          type: number
        isRetained:
          type: boolean
        battingStyle:
          type: ["string", "null"]
        bowlingStyle:
          type: ["string", "null"]
      required: [id, name, role, nationality, isOverseas, priceCrore, isRetained]
...